rename packages, add license comments
