+++ app/src/main/res/menu/menu_conversation_info.xml
@@ -3,7 +3,9 @@
   Nextcloud Android Talk application
 
   @author Tobias Kaminsky
+  @author Marcel Hibbe
   Copyright (C) 2021 Tobias Kaminsky
+  Copyright (C) 2023 Marcel Hibbe (dev@mhibbe.de)
   Copyright (C) 2021 Nextcloud GmbH
  
   This program is free software: you can redistribute it and/or modify
